requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.9.2
httpx==0.27.2
beautifulsoup4==4.12.3
playwright==1.48.0
yt-dlp==2024.11.18
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.32.0
python-dotenv==1.0.1
structlog==24.4.0
python-docx==1.1.2
